Separate file operation and ParserGenerator initialization
The new design reduces the risk of opening a file,
finding that the grammar is invalid, and losing the old contents of
that file due to opening with mode "w". The new design encourages
opening the file as close to the call of generate() as possible.
(build.py and build_v2.py are examples.)